copy: return 0 on success and appropriate errno on failure
This code was still using an old paradigm of returning the
number of bytes associated with a successful submission.
Just return 0 on success instead - if caller needs the
number of bytes for some reason they have the information
to get it.

While here, return an appropriate negated errno where possible -
we especially want ENOMEM returned when an ioat channel runs out
of descriptors.


Reviewed-on: https://review.gerrithub.io/378858
Tested-by: SPDK Automated Test System <sys_sgsw@intel.com>

+++ include/spdk/copy_engine.h
@@ -50,10 +50,10 @@ int spdk_copy_engine_initialize(void);
 int spdk_copy_engine_finish(void);
 
 struct spdk_io_channel *spdk_copy_engine_get_io_channel(void);
-int64_t spdk_copy_submit(struct spdk_copy_task *copy_req, struct spdk_io_channel *ch, void *dst,
-			 void *src, uint64_t nbytes, spdk_copy_completion_cb cb);
-int64_t spdk_copy_submit_fill(struct spdk_copy_task *copy_req, struct spdk_io_channel *ch,
-			      void *dst, uint8_t fill, uint64_t nbytes, spdk_copy_completion_cb cb);
+int spdk_copy_submit(struct spdk_copy_task *copy_req, struct spdk_io_channel *ch, void *dst,
+		     void *src, uint64_t nbytes, spdk_copy_completion_cb cb);
+int spdk_copy_submit_fill(struct spdk_copy_task *copy_req, struct spdk_io_channel *ch,
+			  void *dst, uint8_t fill, uint64_t nbytes, spdk_copy_completion_cb cb);
 size_t spdk_copy_task_size(void);
 
 #endif
